requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.3.1
autopep8==1.5.4
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.3
django-celery-results==2.0.0
django-celery-tasks==0.0.1
django-rest-framework==0.1.0
djangorestframework==3.12.2
flake8==3.8.4
future==0.16.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
kombu==4.6.11
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.4
raven==6.10.0
redis==3.5.3
requests==2.22.0
retrying==1.3.3
sentry-sdk==0.19.4
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
